Fix inconsistency between signing host, and request host
The authorization string was generated using a hardcoded host
generated by `v4.lua`. This differs from the request host which
is the host defined in the API `hosts`. Thus, AWS would return:

	The request signature we calculated does not match the
        signature you provided

This commit fixes the error by using the `Host` defined in the
header, and by setting the upstream host to the one that was used
in signing the request. It essentially forces `preserve_host`.

The `Host` header is set to `ngx.var.host` as it will not contain
the port. The port is passed separately using `ngx.var.port`.
The port will not be used for the signing request, as it will
not be used in the proxied request.

Furthermore, it sets the `Connection` header to `nil`. This ensures
the headers used in the signing request matches the proxied request.
Otherwise, it may be set to `keep-alive` from user -> Kong,
but `close` from Kong -> AWS.

This commit should allow requests to be proxied via URI, and via
Host.